Move verifier into dedicated module
